Move safe_get_lang_filename to languagemanager.py, change fallback extension to .invalid_language
