+++ cms/server/admin/templates/fragments/user_test_row.html
@@ -52,7 +52,7 @@
   </td>
   <td>
     {% for filename, sub_file in ut.files|dictsort(by="key") %}
-      {% set real_filename = get_lang_filename(ut, filename) %}
+      {% set real_filename = get_lang_filename(ut.language, filename) %}
     <a href="javascript:void(0);" onclick="utils.show_file('{{ real_filename }}','{{ url("user_test_file", sub_file.id) }}')">{{ real_filename }}</a><br/>
     {% endfor %}
   </td>
