+++ cms/server/admin/templates/macro/submission.html
@@ -165,7 +165,7 @@ dataset (Dataset|None): the dataset to show results for, or if not defined use
   </td>
   <td>
     {% for filename, sub_file in s.files|dictsort(by="key") %}
-      {% set real_filename = get_lang_filename(s, filename) %}
+      {% set real_filename = get_lang_filename(s.language, filename) %}
     <a href="javascript:void(0);" onclick="utils.show_file('{{ real_filename }}','{{ url("submission_file", sub_file.id) }}')">{{ real_filename }}</a><br/>
     {% endfor %}
   </td>
