+++ cms/server/admin/templates/user_test.html
@@ -41,7 +41,7 @@
         <td>
           {% for filename in ut.task.submission_format %}
             {% if filename in ut.files %}
-              {% set real_filename = get_lang_filename(ut, filename) %}
+              {% set real_filename = get_lang_filename(ut.language, filename) %}
           <a href="javascript:void(0);" onclick="utils.show_file('{{ real_filename }}','{{ url("user_test_file", ut.files[filename].id) }}')">
             {{ real_filename }}
           </a>
